+++ src/main/java/calculator/utils/CalculatorRegex.java
@@ -1,0 +1,8 @@
+package calculator.utils;
+
+public class CalculatorRegex {
+
+    public static final String DEFAULT_SEPERATOR_REGEX = ",|:";
+    public static final String CUSTOM_SEPERATOR_REGEX = "//(.)\\\\n(.*)";
+    public static final String POSITIVE_NUMBER_REGX = "^[0-9]+$";
+}
