+++ src/main/java/calculator/validator/CalculatorValidator.java
@@ -1,0 +1,24 @@
+package calculator.validator;
+
+import calculator.utils.CalculatorRegex;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+public class CalculatorValidator {
+    
+    public static void validateExpression(String expression) {
+
+        // 입력값이 없거나 null인 경우 예외 처리
+        if (expression.isEmpty() || expression == null) {
+            throw new IllegalArgumentException("입력값이 없습니다.");
+        }
+
+        // 커스텀 구분자가 존재하는 경우
+        if (expression.startsWith("//")) {
+            Matcher matcher = Pattern.compile(CalculatorRegex.CUSTOM_SEPERATOR_REGEX).matcher(expression);
+            if (!matcher.find()) {
+                throw new IllegalArgumentException("커스텀 구분자가 잘못되었습니다.");
+            }
+        }
+    }
+}
